requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-environ==0.9.0
django-js-asset==2.0.0
django-mptt==0.14.0
django_debug_toolbar==3.8.1
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stripe==5.2.0
urllib3==1.26.14
